It also pulls in changes required for HIP archive
unbundling i.e. D133705 and D135724.

+++ llvm/lib/CodeGen/SelectionDAG/SelectionDAG.cpp
@@ -662,7 +662,7 @@ static void AddNodeIDOperands(FoldingSetNodeID &ID,
   }
 }
 
-static void AddNodeIDNode(FoldingSetNodeID &ID, unsigned short OpC,
+static void AddNodeIDNode(FoldingSetNodeID &ID, unsigned OpC,
                           SDVTList VTList, ArrayRef<SDValue> OpList) {
   AddNodeIDOpcode(ID, OpC);
   AddNodeIDValueTypes(ID, VTList);
